+++ lib/auto_html/filters/google_video.rb
@@ -1,0 +1,6 @@
+# GoogleVideo is discontinued so don't do anything
+# TODO Remove this filter in auto_html v2.0
+AutoHtml.add_filter(:google_video) do |text, options|
+  text
+end
+